+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.2.0 (2017-01-15)
 
   * Adds temperature units and conversion function.
+  * Adds hydrometer reading correction for temperature.
 
 # 0.1.0 (2017-01-15)